liquidity: add rule and target to liqudity manager
This commit adds a rule interface and a set of targets that the manager
can apply rules to. The logic for the threshold rule added will be added
in followup commits, for for focused review.

This setup allows users to configure the level at which they want to
manage liquidity. These general rules can be customized in future by
allowing users to specify rules per channel or per peer. These special
rules (such as excluding a peer, setting a different threshold etc)
will exclude an entity from the more general rule set by these targets.
Users with highly custom setups that wish to be more hands on will have
the option to disable these generalized targets completely, and just
run with custom targets.