regtest: add demo/quick start regtest environment
To help with development of apps that use the Loop daemon (`loopd`) for
liquidity management, it is often useful to test it in controlled
environments
such as on the Bitcoin `regtest` network.

Lightning Labs provides a stripped down version of the Loop server that
works
only for `regtest` and is only published in compiled, binary form
(currently as
a Docker image).

The `docker-compose.yml` shows an example setup and is accompanied by a
quick
start script that should make getting started very easy.