loopd: always skip migration if loop_sqlite.db file already exists
This commit ensures that we don't even check for the existence of
the loop.db file if loop_sqlite.db already present. This is to ensure
that on systems that may by any trigger restore deleted files we don't
try to re-run migration which would otherwise make loopd unable to start
up properly.