assets+loopd: add scaffolding for the asset deposit manager
This commit extends the deposit package with the `Deposit` type and
a bare-bones `Manager`, along with the structural definition of the
`SQLStore` which together implement the basic structure, run loop and
storage for the deposit manager. It is not yet functional and serves
as a foundation for future commits that will gradually extend its
functionality.